moving beacon gateway things in to beacons gateway folder
