+++ src/gateways/IApiResponse.ts
