+++ src/gateways/beacons/BeaconRequestMapper.test.ts
