+++ src/gateways/beacons/BeaconRequestMapper.ts
