+++ src/gateways/beacons/BeaconResponseMapper.test.ts
