+++ src/gateways/beacons/BeaconResponseMapper.ts
