+++ src/gateways/beacons/IBeaconRequest.ts
